Implement Errata E24 for XSLTC, when the value attribute of xsl:number is NaN, infinite or less than 0.5. More details please refer to bug report XALANJ-1979. Patch is reviewed by Morris Kwan (mkwan@ca.ibm.com)
git-svn-id: https://svn.apache.org/repos/asf/xalan/java/trunk@338111 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/xalan/xsltc/dom/NodeCounter.java
@@ -37,7 +37,7 @@ public abstract class NodeCounter {
 
     protected int _node = END;
     protected int _nodeType = DOM.FIRST_TYPE - 1;
-    protected int _value = Integer.MIN_VALUE;
+    protected double _value = Integer.MIN_VALUE;
 
     public final DOM          _document;
     public final DTMAxisIterator _iterator;
@@ -84,7 +84,7 @@ public abstract class NodeCounter {
      * If the user specified a value attribute, use this instead of 
      * counting nodes.
      */
-    public NodeCounter setValue(int value) {
+    public NodeCounter setValue(double value) {
     _value = value;
     return this;
     }
